requirements.txt
gradio==4.22.0
ipykernel==6.29.0
ipywidgets==8.1.2
pillow==10.2.0
python-dotenv==1.0.1
transformers[torch]==4.39.1

